🚧 feat: 지도 생성 시 authority 옵션 추가, cookie 관련 코드 util로 관리

+++ FE/src/apis/mypage.ts
@@ -1,6 +1,10 @@
 import { MypagePostParams } from '@/types/mypage';
+import { getCookie } from '@/utils/cookie';
 
 export const getMypage = async () => {
+  const accessToken = getCookie('access_token');
+  if (!accessToken) throw new Error('accesToken is undefined');
+
   const response = await fetch(`/mypage`);
   const mypageData = await response.json();
 
@@ -12,9 +16,15 @@ export const getMypage = async () => {
 };
 
 export const postMypage = async (mypageRequestBody: MypagePostParams) => {
+  const accessToken = getCookie('access_token');
+  if (!accessToken) throw new Error('accesToken is undefined');
+
   const response = await fetch(`/mypage`, {
     method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
+    headers: {
+      Authorization: `Barer ${accessToken}`,
+      'Content-Type': 'application/json',
+    },
     body: JSON.stringify(mypageRequestBody),
   });
   const mypageData = await response.json();
@@ -30,12 +40,15 @@ export const patchMypage = async (
   patchId: number,
   mypageRequestBody: MypagePostParams
 ) => {
+  const accessToken = getCookie('access_token');
+  if (!accessToken) throw new Error('accesToken is undefined');
   if (!patchId) throw new Error('id is undefined');
   if (!mypageRequestBody) throw new Error('requestbody is undefined');
 
   const response = await fetch(`/mypage/${patchId}`, {
     method: 'PATCH',
     headers: {
+      Authorization: `Barer ${accessToken}`,
       'Content-Type': 'application/json',
     },
     body: JSON.stringify(mypageRequestBody),
@@ -50,9 +63,12 @@ export const patchMypage = async (
 };
 
 export const deleteMypage = async (deleteId: number) => {
+  const accessToken = getCookie('access_token');
+  if (!accessToken) throw new Error('accesToken is undefined');
   const response = await fetch(`mypage/${deleteId}`, {
     method: 'DELETE',
     headers: {
+      Authorization: `Barer ${accessToken}`,
       'Content-Type': 'application/json',
     },
   });
